requirements.txt
flask==3.0.2
groq==1.1.2
Pillow==10.2.0
gunicorn==21.2.0
